+++ src/ScottPlot5/ScottPlot5/Primitives/Color.cs
@@ -197,9 +197,27 @@ public readonly struct Color
         return new SKColor(Red, Green, Blue, Alpha);
     }
 
+    /// <summary>
+    /// Luminance as a fraction from 0 to 1
+    /// </summary>
+    public float Luminance => ToHSL().l;
+
+    /// <summary>
+    /// Hue as a fraction from 0 to 1
+    /// </summary>
+    public float Hue => ToHSL().h;
+
+    /// <summary>
+    /// Saturation as a fraction from 0 to 1
+    /// </summary>
+    public float Saturation => ToHSL().s;
+
+    /// <summary>
+    /// Hue, Saturation, and Luminance (as fractions from 0 to 1)
+    /// </summary>
     public (float h, float s, float l) ToHSL()
     {
-        // Converter adapted from http://en.wikipedia.org/wiki/HSL_color_space
+        // adapted from http://en.wikipedia.org/wiki/HSL_color_space
         float r = Red / 255f;
         float g = Green / 255f;
         float b = Blue / 255f;
@@ -236,6 +254,9 @@ public readonly struct Color
         return (h, s, l);
     }
 
+    /// <summary>
+    /// Create a Color given Hue, Saturation, and Luminance (as fractions from 0 to 1)
+    /// </summary>
     public static Color FromHSL(float hue, float saturation, float luminosity, float alpha = 1)
     {
         // adapted from Microsoft.Maui.Graphics/Color.cs (MIT license)