+++ CHANGELOG.md
@@ -7,6 +7,9 @@ _Not yet on NuGet..._
 * Palette: Improved behavior of automatic color selection for new plottables added to the plot (#4105)
 * Color: Added overload to allow constructing colors from hex strings (#4170, #4165) @aespitia @KroMignon @kebox7
 * WinUI: Improved support and documentation for applications built using Uno Platform (#4123) @agneszitte
+* Color: Added Hue, Saturation, and Luminance properties
+* Colormap: Added `Colormap.GetColormaps()` to allow iterating over all available colormaps
+* Colormap: Added `Colormap.GetImage()` to generate a gradient image using a given colormap
 
 ## ScottPlot 5.0.37
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-07-29_